Expose generator.
